+++ example/endpoints/send-avatar.ts
@@ -3,7 +3,8 @@ import { fileSendingEndpointsFactory } from "../factories";
 import fs from "fs";
 
 export const sendAvatarEndpoint = fileSendingEndpointsFactory.build({
-  methods: ["get"],
+  method: "get",
+  tags: ["files", "users"],
   input: z.object({
     userId: z
       .string()