+++ example/endpoints/update-user.ts
@@ -4,6 +4,7 @@ import { keyAndTokenAuthenticatedEndpointsFactory } from "../factories";
 export const updateUserEndpoint =
   keyAndTokenAuthenticatedEndpointsFactory.build({
     method: "post",
+    tag: "users",
     description: "example user update endpoint",
     input: withMeta(
       z.object({