+++ src/endpoint.ts
@@ -41,6 +41,7 @@ export abstract class AbstractEndpoint {
   public abstract getNegativeMimeTypes(): string[];
   public abstract getSecurity(): LogicalContainer<Security>;
   public abstract getScopes(): string[];
+  public abstract getTags(): string[];
 }
 
 type EndpointProps<
@@ -50,7 +51,8 @@ type EndpointProps<
   M extends Method,
   POS extends ApiResponse,
   NEG extends ApiResponse,
-  SCO extends string
+  SCO extends string,
+  TAG extends string
 > = {
   middlewares: AnyMiddlewareDef[];
   inputSchema: IN;
@@ -59,8 +61,9 @@ type EndpointProps<
   handler: Handler<z.output<IN>, z.input<OUT>, OPT>;
   resultHandler: ResultHandlerDefinition<POS, NEG>;
   description?: string;
-  scopes?: SCO[];
-} & MethodsDefinition<M>;
+} & ({ scopes?: SCO[] } | { scope?: SCO }) &
+  ({ tags?: TAG[] } | { tag?: TAG }) &
+  MethodsDefinition<M>;
 
 export class Endpoint<
   IN extends IOSchema,
@@ -69,7 +72,8 @@ export class Endpoint<
   M extends Method,
   POS extends ApiResponse,
   NEG extends ApiResponse,
-  SCO extends string
+  SCO extends string,
+  TAG extends string
 > extends AbstractEndpoint {
   protected readonly description?: string;
   protected readonly methods: M[] = [];
@@ -79,7 +83,8 @@ export class Endpoint<
   protected readonly outputSchema: OUT;
   protected readonly handler: Handler<z.output<IN>, z.input<OUT>, OPT>;
   protected readonly resultHandler: ResultHandlerDefinition<POS, NEG>;
-  protected readonly scopes?: SCO[];
+  protected readonly scopes: SCO[];
+  protected readonly tags: TAG[];
 
   constructor({
     middlewares,
@@ -89,9 +94,8 @@ export class Endpoint<
     resultHandler,
     description,
     mimeTypes,
-    scopes,
     ...rest
-  }: EndpointProps<IN, OUT, OPT, M, POS, NEG, SCO>) {
+  }: EndpointProps<IN, OUT, OPT, M, POS, NEG, SCO, TAG>) {
     super();
     this.middlewares = middlewares;
     this.inputSchema = inputSchema;
@@ -100,7 +104,20 @@ export class Endpoint<
     this.handler = handler;
     this.resultHandler = resultHandler;
     this.description = description;
-    this.scopes = scopes;
+    this.scopes = [];
+    this.tags = [];
+    if ("scopes" in rest && rest.scopes) {
+      this.scopes.push(...rest.scopes);
+    }
+    if ("scope" in rest && rest.scope) {
+      this.scopes.push(rest.scope);
+    }
+    if ("tags" in rest && rest.tags) {
+      this.tags.push(...rest.tags);
+    }
+    if ("tag" in rest && rest.tag) {
+      this.tags.push(rest.tag);
+    }
     if ("methods" in rest) {
       this.methods = rest.methods;
     } else {
@@ -153,7 +170,11 @@ export class Endpoint<
   }
 
   public override getScopes(): SCO[] {
-    return this.scopes || [];
+    return this.scopes;
+  }
+
+  public override getTags(): TAG[] {
+    return this.tags;
   }
 
   #getDefaultCorsHeaders(): Record<string, string> {